3rdparty/hpcanalysis/pyproject.toml
# SPDX-FileCopyrightText: Contributors to the HPCToolkit Project
# 
# SPDX-License-Identifier: Apache-2.0

[project]
name = "hpcanalysis"
version = "0.1.0"
description = "Framework for analyzing application performance at exascale"
authors = [
    {name = "Dragana Grbic", email = "draganaurosgrbic@gmail.com"},
]
dependencies = [
    "pandas>=2.0.0,<=2.2.3",
    "tabulate>=0.8.0,<=0.9.0",
    "scikit-learn>=1.3.0,<=1.6.1",
    "joblib>=1.4.0,<=1.5.0",
    "ipython>=8.0.0,<=9.2.0",
    "ipytree>=0.2.0,<=0.2.2",
    "treelib>=1.6.0,<=1.7.1",
    "llnl-hatchet>=2024.1.3,<2024.2",
    "llnl-thicket>=2024.2.1,<2024.3",
    "dask[complete]>=2024.5.0,<=2024.9.0",

    # CUDA 13 / RAPIDS Packages for GPU acceleration
    "libcudf-cu13",
    "pylibcudf-cu13",
    "cudf-cu13",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "Apache-2.0"}
keywords = ["exascale", "performance analysis", "high-performance computing", "HPC", "profiling"]

[build-system]
requires = ["meson-python", "pybind11"]
build-backend = "mesonpy"

[dependency-groups]
dev = [
    "black>=24.0.0,<=25.1.0",
    "isort>=5.12.0,<=6.0.1",
    "pytest>=8.0.0,<=8.3.5",
    "pytest-cov>=5.0.0,<=6.1.1",
    "flake8>=6.0.0,<=7.2.0",
    "mypy>=1.0.0,<=1.15.0",
    "sphinx>=7.0.0,<=8.2.3",
    "sphinx-rtd-theme>=2.0.0,<=3.0.2",
]

[project.optional-dependencies]
jupyter-notebook = [
    "jupyterlab>=3.0.0,<=4.4.2",
    "notebook>=6.0.0,<=7.4.2",
]